fixed compile error

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -8,7 +8,7 @@
 
 hash_table_t *hash_table_create(unsigned long int size)
 {
-	int i = 0;
+	unsigned long int i = 0;
 	hash_table_t *hash_table_ptr = NULL;
 	hash_node_t *head = NULL;
